Refactor out the options on the cases in location page to its own file
