+++ src/styles.scss
@@ -15,7 +15,13 @@ body {
 }
 
 #root {
-  height: 100%;
+  min-height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+.footer-fill {
+  flex-grow: 1
 }
 
 .flex-1 {
